refactor: change name of reducer to settingsReducer
