+++ src/components/SettingsList/ShowHistory/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { toggleShowHistory } from '@actions/control';
+import { toggleShowHistory } from '@actions/settings';
 
 import { Input, Label, Wrapper } from './styled';
 
 const TogglerHistoryFC = () => {
     const dispatch = useDispatch();
-    const checked = useSelector(({ control }) => control.isShowHistory);
+    const checked = useSelector(({ settings }) => settings.isShowHistory);
     const handleShowHistory = () => {
         dispatch(toggleShowHistory());
     };
