+++ src/store/store.jsx
@@ -2,8 +2,8 @@ import { combineReducers, legacy_createStore as createStore } from 'redux';
 import { persistReducer, persistStore } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 
-import controlReducer from './reducers/control';
 import displayReducer from './reducers/display';
+import settingsReducer from './reducers/settings';
 
 const persistConfig = {
     key: 'root',
@@ -12,7 +12,7 @@ const persistConfig = {
 
 const rootReducers = combineReducers({
     display: displayReducer,
-    control: controlReducer,
+    settings: settingsReducer,
 });
 
 const persistedReducer = persistReducer(persistConfig, rootReducers);
